672: Remove the AzureClient interface to use a direct implementation.
